+++ spicedb-go/client/experimental.go
@@ -2,7 +2,6 @@ package client
 
 import (
 	"context"
-	"fmt"
 
 	v1 "github.com/authzed/spicedb-clients/proto-clients/spicedb-go-proto/gen/authzed/api/v1"
 	"github.com/authzed/spicedb-clients/spicedb-go/rel"
@@ -24,7 +23,7 @@ func (c *Client) RegisterRelationshipCounter(ctx context.Context, name string, f
 		RelationshipFilter: f.ToProto(),
 	})
 	if err != nil {
-		return fmt.Errorf("spicedb: register relationship counter: %w", err)
+		return mapGRPCError("register relationship counter", err)
 	}
 	return nil
 }
@@ -38,7 +37,7 @@ func (c *Client) CountRelationships(ctx context.Context, name string) (result *C
 		Name: name,
 	})
 	if err != nil {
-		return nil, false, fmt.Errorf("spicedb: count relationships: %w", err)
+		return nil, false, mapGRPCError("count relationships", err)
 	}
 
 	if resp.GetCounterStillCalculating() {
@@ -60,7 +59,7 @@ func (c *Client) UnregisterRelationshipCounter(ctx context.Context, name string)
 		Name: name,
 	})
 	if err != nil {
-		return fmt.Errorf("spicedb: unregister relationship counter: %w", err)
+		return mapGRPCError("unregister relationship counter", err)
 	}
 	return nil
 }